test(onboard): split sandbox naming coverage
